+++ src/editor/base-editor.ts
@@ -1,6 +1,6 @@
 // lovelace card imports.
 import { css, LitElement } from 'lit';
-import { property } from 'lit/decorators.js';
+import { property, state } from 'lit/decorators.js';
 import { HomeAssistant } from '../types/home-assistant-frontend/home-assistant';
 import { fireEvent } from '../types/home-assistant-frontend/fire-event';
 
@@ -27,7 +27,10 @@ export abstract class BaseEditor extends LitElement {
   @property({ attribute: false }) footerBackgroundColor?: string;
   @property({ attribute: false }) public soundTouchDevice!: ISoundTouchDevice | undefined;
 
-  //@state() protected soundTouchDevice!: ISoundTouchDevice | undefined;
+  @state() public playerImage?: string;
+  @state() public playerMediaContentId?: string;
+  @state() public vibrantImage?: string;
+  @state() public vibrantMediaContentId?: string;
 
   /** MediaPlayer instance created from the configuration entity id. */
   public player!: MediaPlayer;